Backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
    